LukittuNode modularity
Added
-Routes
-Models
-Main.js
-db_pool for pool.js (could be separate)
-Stripped server.js for more readability
